+++ package.json
@@ -16,7 +16,6 @@
     "@next/mdx": "^14.2.3",
     "@prisma/client": "^6.1.0",
     "@types/mdx": "^2.0.13",
-    "better-sqlite3": "^11.7.0",
     "classnames": "^2.3.2",
     "date-fns": "^2.28.0",
     "gray-matter": "^4.0.3",
@@ -25,7 +24,6 @@
     "next": "latest",
     "next-images": "^1.8.5",
     "next-mdx-remote": "^5.0.0",
-    "node-sass": "^7.0.3",
     "react": "^18.2.0",
     "react-burger-menu": "^3.0.9",
     "react-custom-scrollbars-2": "^4.5.0",
@@ -35,9 +33,8 @@
     "rehype-slug": "^6.0.0",
     "remark": "^14.0.2",
     "remark-html": "^15.0.1",
-    "sass": "^1.56.1",
+    "sass": "^1.83.1",
     "sharp": "^0.33.1",
-    "sqlite3": "^5.1.7",
     "tocbot": "^4.25.0",
     "typescript": "^4.7.4"
   },
